requirements.txt
aniso8601==8.0.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.16.0
lxml==4.5.2
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.5
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-file==1.5.1
six==1.15.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==2.2.3
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
